src/main/java/helloworld/GreetingController.java
package helloworld;

import java.util.concurrent.atomic.AtomicLong;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping(value = "/api/v1")
public class GreetingController {
	private static final String template = "Hello, %s!";
    private final AtomicLong counter = new AtomicLong();

    @Value("${spring.application.name}")
	private String applicationName;
    
    @RequestMapping("/greeting")
    public Greeting greeting(@RequestParam(value="name", defaultValue="World") String name) {
    	System.out.println("GreetingController.greeting() - applicationName: " + applicationName);
        return new Greeting(counter.incrementAndGet(),
                            String.format(template, name));
    }
}
